requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.16.0
channels==3.0.4
charset-normalizer==3.1.0
constantly==23.10.4
cryptography==41.0.7
daphne==3.0.2
distlib==0.3.6
Django==4.0
django-cors-headers==4.0.0
exceptiongroup==1.1.2
factory-boy==3.3.0
Faker==22.4.0
filelock==3.12.2
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==2.0.0
multidict==6.0.4
outcome==1.2.0
packaging==23.2
pick==2.2.0
Pillow==10.1.0
platformdirs==3.8.0
pluggy==1.3.0
psutil==5.9.7
pyaes==1.6.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.3.0
pypng==0.20220715.0
PySocks==1.7.1
pytest==7.4.4
pytest-django==4.7.0
python-dateutil==2.8.2
pytz==2023.3
qrcode==7.4.2
requests==2.31.0
rsa==4.9
selenium==4.10.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
telepot==12.7
termcolor==2.3.0
tomli==2.0.1
trio==0.22.2
trio-websocket==0.10.3
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.6.2
tzdata==2023.3
urllib3==2.0.2
virtualenv==20.23.1
wsproto==1.2.0
yarl==1.9.2
zope.interface==6.1
